+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 1.2.2.pre
 -----------------
+  * Add skeeks\sx\helpers\BackendResponseHelper
   * The declaration deprecated
 
 1.2.1